Move SchemaQuery and PropDescType to individual files
The struggle continues.
