+++ packages/components/src/internal/components/base/models/schemas.ts
@@ -18,7 +18,8 @@ import { Query } from '@labkey/api';
 
 // Have to import SchemaQuery like this or SchemaQuery.create fails because SchemaQuery is undefined.
 import { SchemaQuery } from '../../../../public/SchemaQuery';
-import { QueryInfo, SchemaDetails } from '../../../..';
+import { QueryInfo } from '../../../../public/QueryInfo';
+import { SchemaDetails } from '../../base/models/model';
 
 // Created By / Modified By
 export const CBMB = List<string>(['Created', 'CreatedBy', 'Modified', 'ModifiedBy']);
