+++ packages/components/src/internal/components/domainproperties/Lookup/Fields.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { List } from 'immutable';
 import { FormControl } from 'react-bootstrap';
 
-import { decodeLookup, encodeLookup, PropDescType } from '../models';
+import { decodeLookup, encodeLookup } from '../models';
+import { PropDescType } from '../PropDescType';
 
 import { createFormInputName } from '../actions';
 import { DOMAIN_FIELD_LOOKUP_CONTAINER, DOMAIN_FIELD_LOOKUP_QUERY, DOMAIN_FIELD_LOOKUP_SCHEMA } from '../constants';
