+++ packages/components/src/internal/components/domainproperties/list/SetKeyFieldNamePanel.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Alert, Col, FormControl, Row } from 'react-bootstrap';
 import { Utils } from '@labkey/api';
 
-import { AUTOINT_TYPE, DomainField, IAppDomainHeader, IDomainField, PropDescType } from '../models';
+import { DomainField, IAppDomainHeader, IDomainField } from '../models';
 import { createNewDomainField } from '../actions';
 
 import { DomainFieldLabel } from '../DomainFieldLabel';
@@ -10,6 +10,7 @@ import { DomainFieldLabel } from '../DomainFieldLabel';
 import { DOMAIN_FIELD_PRIMARY_KEY_LOCKED } from '../constants';
 
 import { ListModel } from './models';
+import { AUTOINT_TYPE, PropDescType } from '../PropDescType';
 
 const AUTO_INC_KEY_OPTION_TEXT = 'Auto integer key';
 const AUTO_INC_KEY_OPTION_VALUE = -1;
