+++ packages/components/src/internal/components/domainproperties/validation/RangeValidationOptions.tsx
@@ -13,7 +13,8 @@ import {
     DOMAIN_VALIDATOR_REMOVE,
 } from '../constants';
 
-import { PropDescType, PropertyValidator } from '../models';
+import { PropertyValidator } from '../models';
+import { PropDescType } from '../PropDescType';
 
 import { LabelHelpTip } from '../../base/LabelHelpTip';
