+++ packages/components/src/internal/util/utils.spec.ts
@@ -16,8 +16,6 @@
 import { fromJS, List, Map } from 'immutable';
 import { PermissionTypes } from '@labkey/api';
 
-import { SchemaQuery } from '../..';
-
 import {
     TEST_USER_APP_ADMIN,
     TEST_USER_ASSAY_DESIGNER,
@@ -32,7 +30,6 @@ import {
     contains,
     getCommonDataValues,
     getDisambiguatedSelectInputOptions,
-    getSchemaQuery,
     getUpdatedData,
     getUpdatedDataFromGrid,
     hasAllPermissions,
@@ -41,7 +38,6 @@ import {
     naturalSort,
     resolveKey,
     resolveKeyFromJson,
-    resolveSchemaQuery,
     similaritySortFactory,
     toLowerSafe,
     unorderedEqual,
@@ -71,45 +67,6 @@ describe('resolveKeyFromJson', () => {
     test('schema name with multiple parts', () => {
         expect(resolveKeyFromJson({ schemaName: ['one', 'Two', 'thrEE$'], queryName: 'four' })).toBe(
             'one$ptwo$pthree$dd/four'
-        );
-    });
-});
-
-describe('resolveSchemaQuery', () => {
-    test('handle undefined schemaQuery', () => {
-        expect(resolveSchemaQuery(undefined)).toBeNull();
-    });
-
-    test('schema without encoding required', () => {
-        const schemaQuery = new SchemaQuery({
-            schemaName: 'name',
-            queryName: 'my favorite query',
-        });
-        expect(resolveSchemaQuery(schemaQuery)).toBe('name/my favorite query');
-    });
-});
-
-describe('getSchemaQuery', () => {
-    test('no decoding required', () => {
-        const expected = new SchemaQuery({
-            schemaName: 'name',
-            queryName: 'query',
-        });
-        expect(getSchemaQuery('name/query')).toEqual(expected);
-    });
-
-    test('decoding required', () => {
-        expect(getSchemaQuery('my$Sname/just$pask')).toEqual(
-            new SchemaQuery({
-                schemaName: 'my/name',
-                queryName: 'just.ask',
-            })
-        );
-        expect(getSchemaQuery('one$ptwo$pthree$d/q1')).toEqual(
-            new SchemaQuery({
-                schemaName: 'one.two.three$',
-                queryName: 'q1',
-            })
         );
     });
 });
