+++ packages/components/src/internal/util/utils.ts
@@ -16,8 +16,8 @@
 import { List, Map, Set, Iterable } from 'immutable';
 import { Utils } from '@labkey/api';
 
-import { SchemaQuery, User } from '../..';
-import { hasParameter, toggleParameter } from '../..';
+import { User } from '../components/base/models/User';
+import { hasParameter, toggleParameter } from '../url/ActionURL';
 
 const emptyList = List<string>();
 
@@ -67,18 +67,6 @@ export function resolveKeyFromJson(json: { schemaName: string[]; queryName: stri
             .join('.'),
         json.queryName
     );
-}
-
-// TODO: resolveSchemaQuery should have a better name, and it should be added as a property on the SchemaQuery record
-//  class. I'm really not sure what resolve is supposed to mean in this context, but I think we can add this as a
-//  property called "key", or something similar since it mostly seems to be used as a state key.
-export function resolveSchemaQuery(schemaQuery: SchemaQuery): string {
-    return schemaQuery ? resolveKey(schemaQuery.getSchema(), schemaQuery.getQuery()) : null;
-}
-
-export function getSchemaQuery(encodedKey: string): SchemaQuery {
-    const [encodedSchema, encodedQuery] = encodedKey.split('/');
-    return SchemaQuery.create(decodePart(encodedSchema), decodePart(encodedQuery));
 }
 
 /**
